Azure/PythonApp/requirements.txt
flask
azure.cosmos
requests
azure-cognitiveservices-vision-computervision
